2009-08-26	Sebastian Huber <Sebastian.Huber@embedded-brains.de>
	* sapi/include/rtems/config.h: Changed stack allocator signature to
	match the one of _Workspace_Allocate().

+++ cpukit/sapi/include/rtems/config.h
@@ -145,7 +145,7 @@ typedef struct {
   /** The BSP may want to provide it's own stack allocation routines.
    *  In this case, the BSP will provide this stack allocation hook.
    */
-  void *                       (*stack_allocate_hook)( uint32_t );
+  void *                       (*stack_allocate_hook)( size_t );
 
   /** The BSP may want to provide it's own stack free routines.
    *  In this case, the BSP will provide this stack free hook.
